moving shared functons to helper
